refactor: fix findFilteredEvents, add save to stat
