+++ ewm/src/main/java/ru/practicum/ewm/event/AdminUpdateEventRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.request;
+package ru.practicum.ewm.event;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
