+++ ewm/src/main/java/ru/practicum/ewm/event/controller/EventAdminController.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.ewm.event.dto.EventFullDto;
 import ru.practicum.ewm.event.service.EventAdminService;
-import ru.practicum.ewm.request.AdminUpdateEventRequest;
+import ru.practicum.ewm.event.AdminUpdateEventRequest;
 
 import java.util.List;
