+++ ewm/src/main/java/ru/practicum/ewm/event/service/EventAdminService.java
@@ -1,7 +1,7 @@
 package ru.practicum.ewm.event.service;
 
 import ru.practicum.ewm.event.dto.EventFullDto;
-import ru.practicum.ewm.request.AdminUpdateEventRequest;
+import ru.practicum.ewm.event.AdminUpdateEventRequest;
 
 import java.util.List;
