+++ ewm/src/main/java/ru/practicum/ewm/event/service/EventPublicService.java
@@ -16,7 +16,9 @@ public interface EventPublicService {
             Boolean onlyAvailable,
             String sort,
             Integer from,
-            Integer size
+            Integer size,
+            String clientIp,
+            String endpointPath
     );
 
     EventFullDto findFullEventInfo(Long eventId, String clientIp, String endpointPath);
