--- ewm/src/main/java/ru/practicum/ewm/request/NewUserRequest.java
@@ -1,4 +1,0 @@
-package ru.practicum.ewm.request;
-
-public class NewUserRequest {
-}
