+++ ewm/src/main/java/ru/practicum/ewm/user/NewUserRequest.java
@@ -1,0 +1,4 @@
+package ru.practicum.ewm.user;
+
+public class NewUserRequest {
+}
